+++ tests/tests.py
@@ -1,9 +1,10 @@
-from django.contrib import admin
+import uuid
+
 from django.contrib.auth.models import User
 from django.utils.timezone import now
 from django.test import TestCase
-import uuid
-from ordered_model.tests.models import (
+
+from tests.models import (
     Answer,
     Item,
     Question,
@@ -11,10 +12,13 @@ from ordered_model.tests.models import (
     CustomOrderFieldModel,
     Pizza,
     Topping,
-    PizzaToppingsThroughModel
+    PizzaToppingsThroughModel,
+    OpenQuestion,
+    MultipleChoiceQuestion,
+    ItemGroup,
+    GroupedItem,
+    TestUser
 )
-from ordered_model.tests.models import TestUser
-from .admin import ItemAdmin
 
 
 class OrderGenerationTests(TestCase):
@@ -23,6 +27,7 @@ class OrderGenerationTests(TestCase):
         self.assertEqual(first_item.order, 0)
         second_item = Item.objects.create()
         self.assertEqual(second_item.order, 1)
+
 
 class ModelTestCase(TestCase):
     fixtures = ['test_items.json']
@@ -87,6 +92,10 @@ class ModelTestCase(TestCase):
         self.assertNames(['1', '2', '4', '3'])
         Item.objects.get(pk=3).to(1)
         self.assertNames(['1', '3', '2', '4'])
+
+    def test_to_not_int(self):
+        with self.assertRaises(TypeError):
+            Item.objects.get(pk=4).to('1')
 
     def test_top(self):
         Item.objects.get(pk=4).top()
@@ -158,7 +167,14 @@ class OrderWithRespectToTests(TestCase):
 
     def test_swap(self):
         with self.assertRaises(ValueError):
-            self.q1_a1.swap([self.q2_a1])
+            self.q1_a1.swap(self.q2_a1)
+
+        self.q1_a1.swap(self.q1_a2)
+        self.assertSequenceEqual(
+            Answer.objects.values_list('pk', 'order'), [
+            (self.q1_a2.pk, 0), (self.q1_a1.pk, 1),
+            (self.q2_a1.pk, 0), (self.q2_a2.pk, 1)
+        ])
 
     def test_up(self):
         self.q1_a2.up()
@@ -237,7 +253,7 @@ class CustomPKTest(TestCase):
                 (self.item4.pk, 3)
             ]
         )
-    
+
     def test_order_to_extra_update(self):
         modified_time = now()
         self.item1.to(3, extra_update={'modified':modified_time})
@@ -246,12 +262,12 @@ class CustomPKTest(TestCase):
                 (self.item2.pk, 0, modified_time),
                 (self.item3.pk, 1, modified_time),
                 (self.item4.pk, 2, modified_time),
-                # This one is the primary item being operated on and modified would be 
+                # This one is the primary item being operated on and modified would be
                 # handled via auto_now or something
                 (self.item1.pk, 3, None)
             ]
         )
-    
+
     def test_bottom_extra_update(self):
         modified_time = now()
         self.item1.bottom(extra_update={'modified':modified_time})
@@ -260,12 +276,12 @@ class CustomPKTest(TestCase):
                 (self.item2.pk, 0, modified_time),
                 (self.item3.pk, 1, modified_time),
                 (self.item4.pk, 2, modified_time),
-                # This one is the primary item being operated on and modified would be 
+                # This one is the primary item being operated on and modified would be
                 # handled via auto_now or something
                 (self.item1.pk, 3, None)
             ]
         )
-    
+
     def test_top_extra_update(self):
         modified_time = now()
         self.item4.top(extra_update={'modified':modified_time})
@@ -274,12 +290,12 @@ class CustomPKTest(TestCase):
                 (self.item4.pk, 0, None),
                 (self.item1.pk, 1, modified_time),
                 (self.item2.pk, 2, modified_time),
-                # This one is the primary item being operated on and modified would be 
+                # This one is the primary item being operated on and modified would be
                 # handled via auto_now or something
                 (self.item3.pk, 3, modified_time)
             ]
         )
-    
+
     def test_below_extra_update(self):
         modified_time = now()
         self.item1.below(self.item4, extra_update={'modified':modified_time})
@@ -288,12 +304,12 @@ class CustomPKTest(TestCase):
                 (self.item2.pk, 0, modified_time),
                 (self.item3.pk, 1, modified_time),
                 (self.item4.pk, 2, modified_time),
-                # This one is the primary item being operated on and modified would be 
+                # This one is the primary item being operated on and modified would be
                 # handled via auto_now or something
                 (self.item1.pk, 3, None)
             ]
         )
-    
+
     def test_above_extra_update(self):
         modified_time = now()
         self.item4.above(self.item1, extra_update={'modified':modified_time})
@@ -302,12 +318,12 @@ class CustomPKTest(TestCase):
                 (self.item4.pk, 0, None),
                 (self.item1.pk, 1, modified_time),
                 (self.item2.pk, 2, modified_time),
-                # This one is the primary item being operated on and modified would be 
+                # This one is the primary item being operated on and modified would be
                 # handled via auto_now or something
                 (self.item3.pk, 3, modified_time)
             ]
         )
-    
+
     def test_delete_extra_update(self):
         modified_time = now()
         self.item1.delete(extra_update={'modified':modified_time})
@@ -426,10 +442,17 @@ class CustomOrderFieldTest(TestCase):
 
 class OrderedModelAdminTest(TestCase):
     def setUp(self):
-        user = User.objects.create_superuser("admin", "a@example.com", "admin")
+        User.objects.create_superuser("admin", "a@example.com", "admin")
         self.assertTrue(self.client.login(username="admin", password="admin"))
-        item1 = Item.objects.create(name='item1')
-        item2 = Item.objects.create(name='item2')
+        Item.objects.create(name='item1')
+        Item.objects.create(name='item2')
+
+        self.ham = Topping.objects.create(name='Ham')
+        self.pineapple = Topping.objects.create(name='Pineapple')
+
+        self.pizza = Pizza.objects.create(name='Hawaiian Pizza')
+        self.pizza_to_ham = PizzaToppingsThroughModel.objects.create(pizza=self.pizza, topping=self.ham)
+        self.pizza_to_pineapple = PizzaToppingsThroughModel.objects.create(pizza=self.pizza, topping=self.pineapple)
 
     def test_move_up_down_links(self):
         res = self.client.get("/admin/tests/item/")
@@ -452,6 +475,22 @@ class OrderedModelAdminTest(TestCase):
         self.assertRedirects(res, "/admin/tests/item/")
         self.assertEqual(Item.objects.get(name="item1").order, 1)
         self.assertEqual(Item.objects.get(name="item2").order, 0)
+
+    def test_move_up_down_links_ordered_inline(self):
+        res = self.client.get("/admin/tests/pizza/")
+        self.assertEqual(res.status_code, 200)
+        self.assertEqual(self.pizza_to_ham.order, 0)
+        self.assertEqual(self.pizza_to_pineapple.order, 1)
+
+        res = self.client.get("/admin/tests/pizza/{}/pizzatoppingsthroughmodel/{}/move-up/".format(
+            self.pizza.id,
+            self.pineapple.id
+        ), follow=True)
+        self.pizza_to_ham.refresh_from_db()
+        self.pizza_to_pineapple.refresh_from_db()
+        self.assertEqual(self.pizza_to_ham.order, 1)
+        self.assertEqual(self.pizza_to_pineapple.order, 0)
+        self.assertEqual(res.status_code, 200)
 
 
 class OrderWithRespectToTestsManyToMany(TestCase):
@@ -488,7 +527,7 @@ class OrderWithRespectToTestsManyToMany(TestCase):
 
     def test_swap(self):
         with self.assertRaises(ValueError):
-            self.p1_t1.swap([self.p2_t1])
+            self.p1_t1.swap(self.p2_t1)
 
     def test_previous(self):
         self.assertEqual(self.p1_t2.previous(), self.p1_t1)
@@ -588,4 +627,77 @@ class MultiOrderWithRespectToTests(TestCase):
 
     def test_swap_fails(self):
         with self.assertRaises(ValueError):
-            self.q1_u1_a1.swap([self.q2_u1_a2])
+            self.q1_u1_a1.swap(self.q2_u1_a2)
+
+class OrderWithRespectToRelatedModelFieldTests(TestCase):
+    def setUp(self):
+        self.u1 = TestUser.objects.create()
+        self.u2 = TestUser.objects.create()
+        self.u1_g1 = self.u1.item_groups.create()
+        self.u2_g1 = self.u2.item_groups.create()
+        self.u2_g2 = self.u2.item_groups.create()
+        self.u1_g2 = self.u1.item_groups.create()
+        self.u2_g2_i1 = self.u2_g2.items.create()
+        self.u2_g1_i1 = self.u2_g1.items.create()
+        self.u1_g1_i1 = self.u1_g1.items.create()
+        self.u1_g2_i1 = self.u1_g2.items.create()
+
+    def test_saved_order(self):
+        self.assertSequenceEqual(
+            GroupedItem.objects.filter(group__user=self.u1).values_list('pk', 'order'), [
+            (self.u1_g1_i1.pk, 0), (self.u1_g2_i1.pk, 1)
+        ])
+
+        self.assertSequenceEqual(
+            GroupedItem.objects.filter(group__user=self.u2).values_list('pk', 'order'), [
+            (self.u2_g2_i1.pk, 0), (self.u2_g1_i1.pk, 1)
+        ])
+
+    def test_swap(self):
+        i2 = self.u1_g1.items.create()
+        self.assertSequenceEqual(
+            GroupedItem.objects.filter(group__user=self.u1).values_list('pk', 'order'), [
+            (self.u1_g1_i1.pk, 0), (self.u1_g2_i1.pk, 1), (i2.pk, 2)
+        ])
+
+        i2.swap(self.u1_g1_i1)
+        self.assertSequenceEqual(
+            GroupedItem.objects.filter(group__user=self.u1).values_list('pk', 'order'), [
+            (i2.pk, 0), (self.u1_g2_i1.pk, 1), (self.u1_g1_i1.pk, 2)
+        ])
+
+    def test_swap_fails_between_users(self):
+        with self.assertRaises(ValueError):
+            self.u1_g1_i1.swap(self.u2_g1_i1)
+
+    def test_above_between_groups(self):
+        i2 = self.u1_g2.items.create()
+        i2.above(self.u1_g1_i1)
+        self.assertSequenceEqual(
+            GroupedItem.objects.filter(group__user=self.u1).values_list('pk', 'order'), [
+            (i2.pk, 0), (self.u1_g1_i1.pk, 1), (self.u1_g2_i1.pk, 2)
+        ])
+
+class PolymorpicOrderGenerationTests(TestCase):
+    def test_order_of_Baselist(self):
+        o1 = OpenQuestion.objects.create()
+        self.assertEqual(o1.order, 0)
+        o1.save()
+        m1 = MultipleChoiceQuestion.objects.create()
+        self.assertEqual(m1.order, 1)
+        m1.save()
+        m2 = MultipleChoiceQuestion.objects.create()
+        self.assertEqual(m2.order, 2)
+        m2.save()
+        o2 = OpenQuestion.objects.create()
+        self.assertEqual(o2.order, 3)
+        o2.save()
+
+        m2.up()
+        self.assertEqual(m2.order, 1)
+        m1.refresh_from_db()
+        self.assertEqual(m1.order, 2)
+        o2.up()
+        self.assertEqual(o2.order, 2)
+        m1.refresh_from_db()
+        self.assertEqual(m1.order, 3)